Introduce tagfiles provider
Use empty on_move impl for tagfiles.
